# Conflicts:
#	project/common/src/main/java/com/project/domain/Order.java
#	project/order/src/main/java/com/project/controller/OrderController.java
#	project/order/src/main/java/com/project/dao/OrderDao.java
#	project/order/src/main/java/com/project/service/OrderService.java
#	project/project.iml
#	project/user/user.iml

+++ project/common/src/main/java/com/project/domain/Evaluation.java
@@ -3,12 +3,11 @@ package com.project.domain;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.ToString;
-
 @Getter
 @Setter
 @ToString
 public class Evaluation {
-    private Integer point; //评分
+    private Double point; //评分
     private String text; // 评价文字
     private String pic; //图片
     private Integer orderId; //所评价的订单号
